Digit Classifier using Support Vector Machines
train.csv - pixel values in cloumns and digits (0-9) in the rows.

Each image is 28 pixels in height and 28 pixels in width, for a total of 784 pixels in total. Each pixel has a single pixel-value associated with it, indicating the lightness or darkness of that pixel, with higher numbers meaning darker. This pixel-value is an integer between 0 and 255, inclusive.

Task: Create a classifier that predicts the digit using support vecor machines.